Laravel 8.x update

+++ database/migrations/2018_04_12_000003_create_engine_virtual_entity_types_table.php
@@ -27,6 +27,10 @@ class CreateEngineVirtualEntityTypesTable extends Migration
      */
     public function down()
     {
+        if(\App::environment('testing') && config('database.default') == 'sqlite') {
+            return;
+        }
+
         Schema::dropIfExists('engine_virtual_entity_types');
     }
 }
